Move NuFastLBL to a separate file
